// src/types/CoalData.ts
export interface CoalData {
  state: string;
  proved_2021_22: number;
  proved_2022_23: number;
  indicated_2021_22: number;
  indicated_2022_23: number;
  inferred_2021_22: number;
  inferred_2022_23: number;
  total_2021_22: number;
  total_2022_23: number;
  distribution_2021_22: number;
  distribution_2022_23: number;
}